infrastructure/testdriver/freeze.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver freeze method</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
var is_child_frozen = false;
var test = async_test('Test freeze callback.');
window.open('freeze_child.html', 'Child Window');

function set_child_frozen(is_frozen) {
  is_child_frozen = is_frozen;
}

function freeze_child(child_window) {
  test_driver
    .freeze(child_window)
    .then(() => {
      assert_true(is_child_frozen);
      test.done();
    })
    .catch(test.unreached_func("freeze failed"));
}

</script>
